[REFAC: #39] - Juntando todo os erros em um modulo.
close #39
